chore: move proc macro example code to their own files.
