+++ utilities/Cargo.toml
@@ -59,6 +59,7 @@ cf-proc-macros = { workspace = true }
 serde_json = { workspace = true, default-features = true }
 tempfile = { workspace = true }
 reqwest = { workspace = true, features = ["rustls-tls"] }
+derive-where = { workspace = true }
 
 [features]
 std = [
