+++ utilities/src/without_std.rs
@@ -20,6 +20,7 @@ pub mod generate_module;
 pub mod migrations;
 pub mod never;
 pub mod select;
+pub mod tests;
 pub mod type_introspection;
 
 pub const fn bs58_array<const S: usize>(s: &'static str) -> [u8; S] {
